Revert "Move all element related items into a class so that Element is internal and all subs follow."
This reverts commit 69689e79f2342edbcbb6b3f5838a3315379fa0a4.
