+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/forms/Form.kt
@@ -17,10 +17,10 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.asLiveData
 import com.stripe.android.paymentsheet.elements.common.Controller
 import com.stripe.android.paymentsheet.elements.common.DropDown
-import com.stripe.android.paymentsheet.elements.common.Element.FocusRequesterCount
-import com.stripe.android.paymentsheet.elements.common.Element.FormElement
-import com.stripe.android.paymentsheet.elements.common.Element.SectionFieldElementType
+import com.stripe.android.paymentsheet.elements.common.FocusRequesterCount
+import com.stripe.android.paymentsheet.elements.common.FormElement
 import com.stripe.android.paymentsheet.elements.common.Section
+import com.stripe.android.paymentsheet.elements.common.SectionFieldElementType
 import com.stripe.android.paymentsheet.elements.common.TextField
 import com.stripe.android.paymentsheet.specifications.IdentifierSpec
 import com.stripe.android.paymentsheet.specifications.LayoutSpec
